Cleaned up simplify_planes
Cleaned up the simplify_planes function by factoring out a method to clean up
neighboring planes.

Also started on loading a triangle mesh so that it can be simplified.
Started using Eigen.

+++ plane.h
@@ -9,7 +9,7 @@ class Plane;
 struct PointerHash;
 struct DerefCompare;
 
-typedef long plane_id;
+typedef long long plane_id;
 typedef std::unordered_set<std::shared_ptr<Plane>, PointerHash, DerefCompare>
 	plane_set;
 typedef std::pair<std::shared_ptr<Plane>, std::shared_ptr<Plane>> plane_pair;
